+++ projects/alpaca-finance-2.0/index.js
@@ -1,53 +1,11 @@
-const sdk = require("@defillama/sdk");
-const abi = require("./abi.json");
-const { getCache, getConfig } = require("../helper/cache");
-const BigNumber = require("bignumber.js");
+const {lendingTvl, borrowTvl} = require("./moneyMarket");
 
-async function lendingTvl(timestamp, _, _1, {api}) {
-  const balances = {};
-  const moneyMarket = (await getConfig("alpaca-finance-2.0", "https://raw.githubusercontent.com/alpaca-finance/alpaca-v2-money-market/main/.mainnet.json")).moneyMarket;
-
-  const lendingSupplies = (await sdk.api.abi.multiCall({
-    block: api.block,
-    chain: api.chain,
-    abi: abi.getFloatingBalance,
-    calls: moneyMarket.markets.map( (market) => {
-      return {
-        target: moneyMarket.moneyMarketDiamond,
-        params: [market.token]
-      }
-    }),
-  })).output
-
-  for(let i = 0; i < moneyMarket.markets.length; i++) {
-    sdk.util.sumSingleBalance(balances,moneyMarket.markets[i].token, lendingSupplies[i].output, api.chain);
-  }
-
-  return balances;
+async function lendingCal(ts, _, _1, {api}) {
+  return await lendingTvl(api.block, api.chain)
 }
 
-async function borrowTvl(timestamp, _, _1, {api}) {
-  const balances = {};
-  const moneyMarket = (await getConfig("alpaca-finance-2.0", "https://raw.githubusercontent.com/alpaca-finance/alpaca-v2-money-market/main/.mainnet.json")).moneyMarket;
-
-  const borrows = (await sdk.api.abi.multiCall({
-    block: api.block,
-    chain: api.chain,
-    abi: abi.getGlobalDebtValueWithPendingInterest,
-    calls: moneyMarket.markets.map((market) => {
-      return {
-        target: moneyMarket.moneyMarketDiamond,
-        params: [market.token]
-      }
-    })
-  })).output
-
-
-  for(let i = 0; i < moneyMarket.markets.length; i++) {
-    sdk.util.sumSingleBalance(balances, moneyMarket.markets[i].token, borrows[i].output, api.chain);
-  }
-
-  return balances;
+async function borrowCal(ts, _, _1, {api}) {
+  return await borrowTvl(api.block, api.chain)
 }
 
 module.exports = {
@@ -56,7 +14,7 @@ module.exports = {
     doublecounted: false,
     methodology: "Sum floating balance and borrow for each token",
     bsc: {
-      tvl: lendingTvl,
-      borrowed: borrowTvl
+      tvl: lendingCal,
+      borrowed: borrowCal
     },
   };
